Create scheduledScriptToCheckForbiddenDomainsAgainstAllowedURLs.ps1
Script that runs on a log analytics gateway to check Event 105 from the OMS Gateway log in application and service logs and then calls the Add-OMSAllowedHost command to add the denied URL. This is done following a check to see if the domain being added is a match with the allowed URL list. Endswith or equal to check is done before letting a new URL through the LogA GW proxy.